Upload files to 'task4-baseline-modeling'
As I could not push directly from Colab due to permission issues, I uploaded the experiment where I added MADOS data for debris-only classification. The results were not as good as expected. The next step is to integrate the entire MADOS dataset with MARIDA and evaluate the results to see if the performance improves."